Eliminated VarOps namespace (renamed to RAJA)... all that was left was foldl, so the LegacyCompatability.hpp was moved to foldl.hpp  This should eventually be moved to camp?!?
