+++ include/RAJA/policy/MultiPolicy.hpp
@@ -20,10 +20,6 @@
 
 #include "RAJA/config.hpp"
 
-#include <tuple>
-
-#include "RAJA/internal/LegacyCompatibility.hpp"
-
 #include "RAJA/policy/PolicyBase.hpp"
 
 #include "RAJA/internal/get_platform.hpp"
@@ -100,13 +96,13 @@ using policy::multi::MultiPolicy;
 
 namespace detail
 {
-template <size_t... Indices, typename... Policies, typename Selector>
-auto make_multi_policy(VarOps::index_sequence<Indices...>,
+template <camp::idx_t ... Indices, typename... Policies, typename Selector>
+auto make_multi_policy(camp::idx_seq<Indices...>,
                        Selector s,
-                       std::tuple<Policies...> policies)
+                       camp::tuple<Policies...> policies)
     -> MultiPolicy<Selector, Policies...>
 {
-  return MultiPolicy<Selector, Policies...>(s, std::get<Indices>(policies)...);
+  return MultiPolicy<Selector, Policies...>(s, camp::get<Indices>(policies)...);
 }
 }  // namespace detail
 
@@ -134,11 +130,11 @@ auto make_multi_policy(Selector s) -> MultiPolicy<Selector, Policies...>
 /// forall, must return an int in the set 0 to N-1 selecting the policy to use
 /// \return A MultiPolicy containing the given selector s
 template <typename... Policies, typename Selector>
-auto make_multi_policy(std::tuple<Policies...> policies, Selector s)
+auto make_multi_policy(camp::tuple<Policies...> policies, Selector s)
     -> MultiPolicy<Selector, Policies...>
 {
   return detail::make_multi_policy(
-      VarOps::make_index_sequence<sizeof...(Policies)>{}, s, policies);
+      camp::make_idx_seq_t<sizeof...(Policies)>{}, s, policies);
 }
 
 namespace detail
