+++ include/RAJA/policy/cuda/kernel/CudaKernel.hpp
@@ -38,8 +38,6 @@
 #include "RAJA/policy/cuda/MemUtils_CUDA.hpp"
 #include "RAJA/policy/cuda/policy.hpp"
 
-#include "RAJA/internal/LegacyCompatibility.hpp"
-
 #include "RAJA/policy/cuda/kernel/internal.hpp"
 
 namespace RAJA
