+++ include/RAJA/policy/cuda/kernel/internal.hpp
@@ -36,8 +36,6 @@
 #include "RAJA/policy/cuda/MemUtils_CUDA.hpp"
 #include "RAJA/policy/cuda/policy.hpp"
 
-#include "RAJA/internal/LegacyCompatibility.hpp"
-
 
 namespace RAJA
 {
