+++ include/RAJA/policy/hip/kernel/internal.hpp
@@ -37,8 +37,6 @@
 #include "RAJA/policy/hip/MemUtils_HIP.hpp"
 #include "RAJA/policy/hip/policy.hpp"
 
-#include "RAJA/internal/LegacyCompatibility.hpp"
-
 
 namespace RAJA
 {
