+++ include/RAJA/policy/openmp/kernel/OmpSyncThreads.hpp
@@ -22,12 +22,6 @@
 
 #if defined(RAJA_ENABLE_OPENMP)
 
-#include <cassert>
-#include <climits>
-
-#include "RAJA/pattern/detail/privatizer.hpp"
-
-#include "RAJA/pattern/kernel/Collapse.hpp"
 #include "RAJA/pattern/kernel/internal.hpp"
 
 #include "RAJA/util/macros.hpp"
@@ -35,7 +29,6 @@
 
 #include "RAJA/policy/openmp/policy.hpp"
 
-#include "RAJA/internal/LegacyCompatibility.hpp"
 
 
 namespace RAJA
